ci: merge-oss needs sudo
Install sudo, used for the install-circleci-local-cli command.

We could modify the installation script to avoid sudo if it's alrady
running as root, but decided to simply install sudo to avoid the
friction of updating multiple repositories and merges.

Also, moved FAIL_REASON setting to be first step, so we always get a
message even if checkout fails.

[ci skip]