multiregion: job registration fixes (#319)
* copy job from request before mutating

* use JobModifyIndex to prevent concurrent updates

When we query for the current job Version, we'll also get the JobModifyIndex
and use that when we make the `Job.Register` RPC for each region (similar to
using `nomad job run -check-index`). That will prevent us from getting
mismatched of the job versions during concurrent updates, as whichever
register attempt "wins" for a region will cause the other to fail.

Note that it's possible for concurrent updates to both fail in this scenario
and have to be retried.