deploymentwatcher: no extra eval for success/(un)blocking deploys (#341)
The MRD checkpoint sends a spurious eval when deployments succesfully
transition out of running. This extra eval causes a miscalculation in the
reconcile loop for jobs with canaries, forcing the removal of a number of
allocs equal to the number of canaries and leaving the job with less than the
desired total.