MRD: deregister regions that are dropped on update (#376)
When a multi-region job has an update that drops a region, we currently orphan
the region and leave it running. Instead, we want to deregister the unwanted
region. We do so once we know that all regions have their copy of the job
registered and that we've kicked off the deployment. This keeps dropping a
region close in semantics to how we drop task groups in single-region
deployments, and matches expectations given to us in customer feedback.