consul: plubming for specifying consul namespace in job/group
This PR adds the common OSS changes and the ENT changes for adding support
for Consul Namespaces, which is going to be a Nomad Enterprise feature.

Wheras the majority of the plumbing exists in the OSS branch, this PR adds
consul_ent.go, which enables the enterprise functionality. Now, job submitters
will be able to specify `-consul-namespace` on job run, as well as `group.namespace`
in a job specification. Doing so places services and checks in the specified
Consul namespace, and Consul KV reads will come from that namespace if `template`
is used by a task.