quotas: avoid double credit for running but server-terminal allocs
When an allocation update records the transition to client-terminal status,
the state store will credit the quota for that allocation's resources. But the
scheduler worker may not have this information and include the allocation in
the list of stopped allocs. This results in double credit and quota
violations. Only subtract the resources for allocations in the `NodeUpdate`
stops if they are not client-terminal.