deps: update go-eventlogger
Pulls in hashicorp/go-eventlogger#51

Requires not only updating a lot number of transitive dependencies, but
also removing our pinning to protobuf v1.3.4 and dbus
v5.0.1+incompatible. I'm not sure why we were pinned to these versions,
but the protobuf pin was added in hashicorp/nomad#8693 which confusingly
also updated the protobuf requirement to v1.4.2.

Removing both of these pins fixes the build.