command: fix tests for MRD deployments for new exit codes
The monitor for `job run` now returns a non-zero exit code if the
deployment is blocked. This was merged in OSS but impacts the command
package tests that exercise multi-region deployments as well, because
these tests intentionally create a deployment that always ends up in
the `blocked` state. Update the tests to match the new behavior.

References: https://github.com/hashicorp/nomad/pull/11550