quota: fix iterator error when ranked node fails later in stack. (#672)
The quota iterator is responsible for checking whether a planned
allocation breaks the resource quota limits it must adhere to.

In the previous behaviour, a call to Next(), would modify the
iterator stored proposed limit state when calculating whether the
new alloc would exceed quota. The problem here arises when a node
passes through the quota iterator, but fails further down the
stack process. In this scenario, a new node is passed through the
stack. When this new node reaches the quota iterator, a new
proposed quota calculation is performed which includes addition
from both node iterations.

The proposed quota state within the iterator can become inflated
meaning job plans and registrations fail due to quota violations
in error. This is only resolved once the allocation has been
placed on a node and the Reset() iterator function is called.

The change means the iterator always works on a copy of the
proposed quota change.