ci: create temporary workflow stub for merge-oss
GHA only enables a workflow to be manually triggered if it exists
on the default [main] branch.

https://github.community/t/cant-trigger-workflow-manually/121740/53

Moving the merge-oss script is going to take some iteration attempts
on a feature branch because it's kinda complicated. To make this
possible, create a stub for the workflow on the main branch so that
I can actually trigger things while working on the feature branch.