enforce secure variables quota limits
The `quota_usage` table used to track allocation usage has to be recalculated
every time we assign a quota to another namespace. For secure variables this
means iterating over a potentially very large set of variables, so we track
secure variables quota usage on a per-namespace basis in its own table. This
means we need to renormalize this data on read, being careful to copy the
`QuotaUsage` struct we get out of the state store so as not to accidentally
mutate it.