quotas: fix tests
**TestFSM_SnapshotRestore_QuotaUsage**

I don't think this test was working the way it was intended before:

The *intention* seems to be to insert a mock quota with mock usage (as
in: >0 resources used) and ensure that those objects were properly
roundtripped through a statestore snapshot/restore. Seems
straightforward and normal.

...except we reconcile quota usage on upsert (in this and many other
cases). Since there aren't any allocs in this test, the usage is zeroed
out. Before this PR that means the original pointer to the quota usage
*also* saw the zero values produced by reconciliation prior to
insertion! So we inserted 0, wrote 0, restored 0, and compared 0 == 0.

*After this PR* we inserted 0, wrote 0, restored 0, but compared the
original nonzero mock usage values to 0! :boom:

There are a bunch of ways we could fix this. Inserting allocs to create
"actual" mock usage would be one. Instead I just went for retrieving the
upserted quota usage *prior* to snapshot/restore roundtripping and
asserting that the pre-snapshot and post-restore objects are the equal.

**TestStateStore_UpsertQuotaUsage**

Similar issue to above but even more straightforward without the
snapshot/restore confusing things:

1. QuotaUsage upserted
2. QuotaUsage reconciled to 0 during upsertion
3. QuotaUsage read and compared

Prior to this PR the comparison was comparing a pointer to itself! The
original pointer inserted is the same pointer pulled back out of the
statestore, so of course they were equivalent regardless of
reconciliation of field values.

I could have fixed this as above: by expecting 0 instead of the original
mock value. However this seemed like an easy test to switch to calculate
usage based on actual (mock) alloc resources.

I also made the (currently test only) StateStore.UpsertQuotaUsage Copy
incoming usages just like every other statestore place that mutates
them. That will hopefully help us test more realistic usage patterns
instead of accidentally sharing a pointer and comparing it to itself.