Ci/oss merge release/1.2.x 20220913164129 (#857)
* test: fix concurrent map access in `TestStatsFetcher` (#14502)

The map of in-flight RPCs gets cleared by a goroutine in the test without first
locking it to make sure that it's not being accessed concurrently by the stats
fetcher itself. This can cause a panic in tests.

Backport of #14496

* Backport of core: Add client scheduling eligibility to heartbeat into release/1.2.x (#14512)

This pull request was automerged via backport-assistant

* CSI: failed allocation should not block its own controller unpublish (#14484) (#14507)

A Nomad user reported problems with CSI volumes associated with failed
allocations, where the Nomad server did not send a controller unpublish RPC.

The controller unpublish is skipped if other non-terminal allocations on the
same node claim the volume. The check has a bug where the allocation belonging
to the claim being freed was included in the check incorrectly. During a normal
allocation stop for job stop or a new version of the job, the allocation is
terminal. But allocations that fail are not yet marked terminal at the point in
time when the client sends the unpublish RPC to the server.

For CSI plugins that support controller attach/detach, this means that the
controller will not be able to detach the volume from the allocation's host and
the replacement claim will fail until a GC is run. This changeset fixes the
conditional so that the claim's own allocation is not included, and makes the
logic easier to read. Include a test case covering this path.

Also includes two minor extra bugfixes:

* Entities we get from the state store should always be copied before
altering. Ensure that we copy the volume in the top-level unpublish workflow
before handing off to the steps.

* The list stub object for volumes in `nomad/structs` did not match the stub
object in `api`. The `api` package also did not include the current
readers/writers fields that are expected by the UI. True up the two objects and
add the previously undocumented fields to the docs.

Co-authored-by: Tim Gross <tgross@hashicorp.com>

* Backport client: fix data races in config handling (#14139) to 1.2.x (#14536)

* client: fix data races in config handling (#14139)

Before this change, Client had 2 copies of the config object: config and configCopy. There was no guidance around which to use where (other than configCopy's comment to pass it to alloc runners), both are shared among goroutines and mutated in data racy ways. At least at one point I think the idea was to have `config` be mutable and then grab a lock to overwrite `configCopy`'s pointer atomically. This would have allowed alloc runners to read their config copies in data race safe ways, but this isn't how the current implementation worked.

This change takes the following approach to safely handling configs in the client:

1. `Client.config` is the only copy of the config and all access must go through the `Client.configLock` mutex
2. Since the mutex *only protects the config pointer itself and not fields inside the Config struct:* all config mutation must be done on a *copy* of the config, and then Client's config pointer is overwritten while the mutex is acquired. Alloc runners and other goroutines with the old config pointer will not see config updates.
3. Deep copying is implemented on the Config struct to satisfy the previous approach. The TLS Keyloader is an exception because it has its own internal locking to support mutating in place. An unfortunate complication but one I couldn't find a way to untangle in a timely fashion.
4. To facilitate deep copying I made an *internally backward incompatible API change:* our `helper/funcs` used to turn containers (slices and maps) with 0 elements into nils. This probably saves a few memory allocations but makes it very easy to cause panics. Since my new config handling approach uses more copying, it became very difficult to ensure all code that used containers on configs could handle nils properly. Since this code has caused panics in the past, I fixed it: nil containers are copied as nil, but 0-element containers properly return a new 0-element container. No more "downgrading to nil!"

* Remove improperly merged fields

* job_endpoint: check spec for all regions (#14519) (#14544)

* job_endpoint: check spec for all regions

Co-authored-by: Derek Strickland <1111455+DerekStrickland@users.noreply.github.com>

* Backport of docs: Fixing heading order, adding text for links in /docs/ecosystem into release/1.2.x (#14564)

This pull request was automerged via backport-assistant

* Backport of core: move LicenseConfig.Copy to shared file into release/1.2.x (#14568)

This pull request was automerged via backport-assistant

* remove LicenseConfig from ent side

* Backport of multiregion: Check spec changed in all regions (#829)

* job_endpoint: check spec for all regions

Co-authored-by: Tim Gross <tgross@hashicorp.com>
Co-authored-by: hc-github-team-nomad-core <82989552+hc-github-team-nomad-core@users.noreply.github.com>
Co-authored-by: Derek Strickland <1111455+DerekStrickland@users.noreply.github.com>
Co-authored-by: Derek Strickland <derek.strickland@gmail.com>