Ci/oss merge release/1.3.x 20220913164031 (#856)
* backport of commit 75b30c22100af8956f27b9f388d8fa6a2250db02 (#14501)

This pull request was automerged via backport-assistant

* test: fix concurrent map access in `TestStatsFetcher` (#14499)

The map of in-flight RPCs gets cleared by a goroutine in the test without first
locking it to make sure that it's not being accessed concurrently by the stats
fetcher itself. This can cause a panic in tests.

Backport of #14496

* Backport of CSI: failed allocation should not block its own controller unpublish into release/1.3.x (#14508)

This pull request was automerged via backport-assistant

* Backport of core: Add client scheduling eligibility to heartbeat into release/1.3.x (#14511)

This pull request was automerged via backport-assistant

* Backport of job_endpoint: check spec for all regions into release/1.3.x (#14545)

This pull request was automerged via backport-assistant

* client: fix data races in config handling (#14139) (#14535)

Before this change, Client had 2 copies of the config object: config and configCopy. There was no guidance around which to use where (other than configCopy's comment to pass it to alloc runners), both are shared among goroutines and mutated in data racy ways. At least at one point I think the idea was to have `config` be mutable and then grab a lock to overwrite `configCopy`'s pointer atomically. This would have allowed alloc runners to read their config copies in data race safe ways, but this isn't how the current implementation worked.

This change takes the following approach to safely handling configs in the client:

1. `Client.config` is the only copy of the config and all access must go through the `Client.configLock` mutex
2. Since the mutex *only protects the config pointer itself and not fields inside the Config struct:* all config mutation must be done on a *copy* of the config, and then Client's config pointer is overwritten while the mutex is acquired. Alloc runners and other goroutines with the old config pointer will not see config updates.
3. Deep copying is implemented on the Config struct to satisfy the previous approach. The TLS Keyloader is an exception because it has its own internal locking to support mutating in place. An unfortunate complication but one I couldn't find a way to untangle in a timely fashion.
4. To facilitate deep copying I made an *internally backward incompatible API change:* our `helper/funcs` used to turn containers (slices and maps) with 0 elements into nils. This probably saves a few memory allocations but makes it very easy to cause panics. Since my new config handling approach uses more copying, it became very difficult to ensure all code that used containers on configs could handle nils properly. Since this code has caused panics in the past, I fixed it: nil containers are copied as nil, but 0-element containers properly return a new 0-element container. No more "downgrading to nil!"

* Backport of docs: Fixing heading order, adding text for links in /docs/ecosystem into release/1.3.x (#14565)

This pull request was automerged via backport-assistant

* Backport of core: move LicenseConfig.Copy to shared file into release/1.3.x (#14567)

This pull request was automerged via backport-assistant

* remove LicenseConfig from ent side

Co-authored-by: hc-github-team-nomad-core <82989552+hc-github-team-nomad-core@users.noreply.github.com>
Co-authored-by: Tim Gross <tgross@hashicorp.com>