admin: Ci/oss merge main 20230413040125 (#1078)
* acl: bump JWT auth gate to 1.5.4 (#16838)

* api: enable support for setting original job source (#16763)

* api: enable support for setting original source alongside job

This PR adds support for setting job source material along with
the registration of a job.

This includes a new HTTP endpoint and a new RPC endpoint for
making queries for the original source of a job. The
HTTP endpoint is /v1/job/<id>/submission?version=<version> and
the RPC method is Job.GetJobSubmission.

The job source (if submitted, and doing so is always optional), is
stored in the job_submission memdb table, separately from the
actual job. This way we do not incur overhead of reading the large
string field throughout normal job operations.

The server config now includes job_max_source_size for configuring
the maximum size the job source may be, before the server simply
drops the source material. This should help prevent Bad Things from
happening when huge jobs are submitted. If the value is set to 0,
all job source material will be dropped.

* api: avoid writing var content to disk for parsing

* api: move submission validation into RPC layer

* api: return an error if updating a job submission without namespace or job id

* api: be exact about the job index we associate a submission with (modify)

* api: reword api docs scheduling

* api: prune all but the last 6 job submissions

* api: protect against nil job submission in job validation

* api: set max job source size in test server

* api: fixups from pr

* consul/connect: fixed a bug where restarting proxy tasks failed. (#16815)

The first start of a Consul Connect proxy sidecar triggers a run
of the envoy_version hook which modifies the task config image
entry. The modification takes into account a number of factors to
correctly populate this. Importantly, once the hook has run, it
marks itself as done so the taskrunner will not execute it again.

When the client receives a non-destructive update for the
allocation which the proxy sidecar is a member of, it will update
and overwrite the task definition within the taskerunner. In doing
so it overwrite the modification performed by the hook. If the
allocation is restarted, the envoy_version hook will be skipped as
it previously marked itself as done, and therefore the sidecar
config image is incorrect and causes a driver error.

The fix removes the hook in marking itself as done to the view of
the taskrunner.

* drainer: fix codec race condition in integration test (#16845)

msgpackrpc codec handles are specific to a connection and cannot be shared
between goroutines; this can cause corrupted decoding. Fix the drainer
integration test so that we create separate codecs for the goroutines that the
test helper spins up to simulate client updates.

This changeset also refactors the drainer integration test to bring it up to
current idioms and library usages, make assertions more clear, and reduce
duplication.

* docs: document signal handling (#16835)

Expand documentation about Nomad's signal handling behaviors, including removing
incorrect information about graceful client shutdowns.

* deps: bump build and linter tool versions (#16858)

This just helps keep up with the podman driver repo.

* Deployment Status Command Does Not Respect -namespace Wildcard (#16792)

* func: add namespace support for list deployment

* func: add wildcard to namespace filter for deployments

* Update deployment_endpoint.go

* style: use must instead of require or asseert

* style: rename paginator to avoid clash with import

* style: add changelog entry

* fix: add missing parameter for upsert jobs

* core: ensure all Server receiver names are consistent. (#16859)

* e2e: add e2e tests for job submission api (#16841)

* e2e: add e2e tests for job submission api

* e2e: fixup callers of AllocLogs

* fix typo

* deps: update docker to 23.0.3 (#16862)

* [no ci] deps: update docker to 23.0.3

This PR brings our docker/docker dependency (which is hosted at github.com/moby/moby)
up to 23.0.3 (forward about 2 years). Refactored our use of docker/libnetwork to
reference the package in its new home, which is docker/docker/libnetwork (it is
no longer an independent repository). Some minor nearby test case cleanup as well.

* add cl

* E2E: clarify drain `-deadline` and `-force` flag behaviors (#16868)

The `-deadline` and `-force` flag for the `nomad node drain` command only cause
the draining to ignore the `migrate` block's healthy deadline, max parallel,
etc. These flags don't have anything to do with the `kill_timeout` or
`shutdown_delay` options of the jobspec.

This changeset fixes the skipped E2E tests so that they validate the intended
behavior, and updates the docs for more clarity.

* fix: typo (#16873)

* docs: add node meta command docs (#16828)

* docs: add node meta command docs

Fixes #16758

* it helps if you actually add the files to git

* fix typos and examples vs usage

* fix: new state.UpsertJob signature for ent tests

---------

Co-authored-by: Seth Hoenig <shoenig@duck.com>
Co-authored-by: James Rasell <jrasell@users.noreply.github.com>
Co-authored-by: Tim Gross <tgross@hashicorp.com>
Co-authored-by: Juana De La Cuesta <juanita.delacuestamorales@hashicorp.com>
Co-authored-by: Shawn <18649508+NOBLES5E@users.noreply.github.com>
Co-authored-by: Michael Schurter <mschurter@hashicorp.com>