ENT versions of `GetConsulClusterName`
The `GetConsulClusterName` method for tasks and services always returns the
default Consul cluster for Nomad CE. In Nomad ENT, we need to walk up the job
until find a valid cluster, or return the default cluster.

Note that we can't do a namespace lookup to get the default Consul/Vault cluster
because that would cause in-place updates of allocations to switch to a new
Consul/Vault cluster when the namespace is changed, and not when the job is
submitted.

So that means that so long as we do the mutation on job submit, then we can
safely fall back to "default" if it's unset by the time we get to the client,
because that means we're dealing with a legacy job and it hasn't been mutated to
change to a non-default cluster.