numa: add support for numa aware scheduling  (#1292)
* core: plumbing to support numa aware scheduling (#18681)

* core: plumbing to support numa aware scheduling

* core: apply node resources compatibility upon fsm rstore

Handle the case where an upgraded server dequeus an evaluation before
a client triggers a new fingerprint - which would be needed to cause
the compatibility fix to run. By running the compat fix on restore the
server will immediately have the compatible pseudo topology to use.

* lint: learn how to spell pseudo

* lang: add a helper for iterating a map in order

In some cases it is helpful to iterate a map in the sorted order of
the maps keyset - particularly in implementations of some function for
which the tests cannot be deterministic without order.

* numa: enable numa aware scheduling for enterprise

This PR implements the "none", "prefer", and "require" numa affinity
preferences for Nomad Enterprise. Injects constraints into jobspecs
where numa capabilities are required.

none - Nomad will attempt to fill smallest core fragments of each
numa node. The idea is to make use of cores that otherwise are not
likely to be useful for jobs requiring large numa fragments.

prefer - Nomad will compute a set of cores that minimizes the total
numa distance cost, but will not require cores to be on the same node
or even the same socket. Algorithmically this will tend to prefer using
the largest fragment available and spilling over to ther numa nodes as
needed, starting with cores on a node of the same socket.

require - Nomad will use the smallest available set of cores on a
single numa node. If no such set of cores is available on a single
numa node the node will be considered exhasusted on the numa-cores
dimension.

* numa: fixup some docs

* numa: ent hook should ensure cores is set if affinity is not none