E2E: add Consul and Vault license variables to workflow (#1306)
Nomad, Consul, and Vault ENT licenses are in the GH repo secrets because we
can't use the CRT Vault for self-hosted runners, which we need for tests that
deploy real workloads with Nomad running as root. Make sure these are exposed to
the `test-e2e` workflow and document where they come from and when they expire
in the workflow file for the next time a developer is looking for them.